pyproject.toml
[tool.poetry]
name = "EXPLORE"
version = "0.0.0"
description = ""
authors = ["Eckart Cobo-Briesewitz"]
license = "MIT"
readme = "README.md"
packages = [{ include = "explore" }]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
robotic = "0.3.3"
mujoco = "3.3.5"
tqdm = "4.67.1"
gym = "0.26.2"
shimmy = "^2.0"
h5py = "3.14.0"
stable_baselines3 = "2.7.0"
scikit-learn = "1.7.1"
numpy = "*"
matplotlib = "*"
wandb = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
